+++ src/konpy/utils.clj
@@ -16,6 +16,9 @@
 (defn user [request]
   (get-in request [:session :identity]))
 
+(defn admin? [user]
+  (= user (env :admin)))
+
 (def start-day (jt/local-date 2025 4 2))
 
 (defn weeks
@@ -31,15 +34,11 @@
   :rcf)
 
 ; https://groups.google.com/g/clojure/c/Kpf01CX_ClM
-(defn create-hash
-  [data-barray]
-  (.digest (java.security.MessageDigest/getInstance "SHA1") data-barray))
-
 (defn sha1 [s]
-  (-> s
-      (.getBytes "UTF-8")
-      create-hash
-      java.math.BigInteger.))
+  (->> (.getBytes s "UTF-8")
+       (.digest (java.security.MessageDigest/getInstance "SHA1"))
+       (java.math.BigInteger. 1)
+       (format "%x")))
 
 (defn remove-spaces [s]
   (-> s