+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 - compile routes.
 - (set! *default-data-reader-fn* tagged-literal)
 - konpy.db wrapper?
+- when eid=-1, create. otherwise update.
+
+## 0.6.3 (2025-05-06)
+
+- can create tasks.
+- can update tasks.
+- renamed `create-tasks!` to `upsert-tasks!`.
+
+## 0.6.2 (2025-05-05)
+
+- updated libraries
+
+| :file    | :name                       | :current | :latest |
+|----------+-----------------------------+----------+---------|
+| deps.edn | datascript/datascript       | 1.7.4    | 1.7.5   |
+|          | io.github.tonsky/clj-reload | 0.9.4    | 0.9.5   |
+|          | io.github.tonsky/clj-reload | 0.9.4    | 0.9.5   |
+|          | io.github.tonsky/fast-edn   | 1.1.2    | 1.1.3   |
+| pom.xml  | org.clojure/clojure         | 0.6.1    | 1.12.0  |
+
 
 ## 0.6.1 (2025-05-05)